Project 3/Hemit/requirements.txt
# Core dependencies
langchain
langchain-community
langchain-huggingface
langchain-ollama
openai

# Vector database
weaviate-client
langchain-weaviate

# API Server
fastapi
uvicorn[standard]
pydantic

# Web interface
streamlit

# Utilities
python-dotenv
tqdm
sentence-transformers
requests
